refactor SyncDbKeyStoreToDisk to ImportDatabaseKeys
uses go-eth keystore import
